Revert "Merge pull request #30568 from teskje/storage-wallclock-metrics"
This reverts commit 6c0ebf56dc5d6a355c1dc01ee5f8c8142b3bfecb, reversing
changes made to 44b073bc081fa12dbfcd91d9bc8458fef18f7687.
