+++ src/compute-client/src/controller.rs
@@ -38,7 +38,6 @@ use futures::stream::FuturesUnordered;
 use futures::{FutureExt, StreamExt};
 use mz_build_info::BuildInfo;
 use mz_cluster_client::client::ClusterReplicaLocation;
-use mz_cluster_client::metrics::ControllerMetrics;
 use mz_cluster_client::{ReplicaId, WallclockLagFn};
 use mz_compute_types::dataflows::DataflowDescription;
 use mz_compute_types::dyncfgs::COMPUTE_REPLICA_EXPIRATION_OFFSET;
@@ -240,8 +239,7 @@ impl<T: ComputeControllerTimestamp> ComputeController<T> {
         storage_collections: StorageCollections<T>,
         envd_epoch: NonZeroI64,
         read_only: bool,
-        metrics_registry: &MetricsRegistry,
-        controller_metrics: ControllerMetrics,
+        metrics_registry: MetricsRegistry,
         now: NowFn,
         wallclock_lag: WallclockLagFn<T>,
     ) -> Self {
@@ -293,8 +291,6 @@ impl<T: ComputeControllerTimestamp> ComputeController<T> {
             }
         });
 
-        let metrics = ComputeControllerMetrics::new(metrics_registry, controller_metrics);
-
         Self {
             instances: BTreeMap::new(),
             instance_workload_classes,
@@ -306,7 +302,7 @@ impl<T: ComputeControllerTimestamp> ComputeController<T> {
             arrangement_exert_proportionality: 16,
             stashed_response: None,
             envd_epoch,
-            metrics,
+            metrics: ComputeControllerMetrics::new(metrics_registry),
             now,
             wallclock_lag,
             dyncfg: Arc::new(mz_dyncfgs::all_dyncfgs()),
